Added the possibility to calculate  Q_{PM}
Q_{PM} and \Delta Q_{PM} are calculated based on their definition  provided by Sprain et al. (2019, EPSL)
In the current preliminary version the input is assumed to be a numpy binary file containing the time sequence and the 120 Gauss coefficients of degree up to 10 ordered in the standard geomagnetic way.